Refactor Tiptap component import and remove unused toolbar component
