+++ app/[lang]/(unprotected)/travels/components/Toolbar.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import { type Editor } from "@tiptap/react";
-import { Bold, Heading2 } from "lucide-react";
 
 type Props = {
   editor: Editor | null;
@@ -13,14 +12,14 @@ const Toolbar = ({ editor }: Props) => {
   }
 
   return (
-    <div className="border-slate-300 bg-white sticky top-3 left-0 right-0 z-10 flex flex-wrap gap-0.5 rounded-lg border p-5">
+    <div className="">
       <button
         onClick={() => editor.chain().focus().toggleHeading({ level: 2 }).run()}
         className={`editor-btn font-extrabold ${
           editor.isActive("heading", { level: 2 }) && "active-editor-btn"
         }`}
       >
-        H2
+        Overskrift
       </button>
     </div>
   );
