+++ package.json
@@ -15,6 +15,7 @@
     "@hookform/resolvers": "^3.1.1",
     "@tailwindcss/forms": "^0.5.4",
     "@tailwindcss/line-clamp": "^0.4.4",
+    "@tiptap/extension-heading": "^2.1.13",
     "@tiptap/pm": "^2.1.12",
     "@tiptap/react": "^2.1.12",
     "@tiptap/starter-kit": "^2.1.12",
