+++ tailwind.config.js
@@ -8,7 +8,7 @@ module.exports = {
     // Or if using `src` directory:
     "./src/**/*.{js,ts,jsx,tsx}",
   ],
-  plugins: [require("@tailwindcss/line-clamp"), require('@tailwindcss/forms'),],
+  plugins: [require('@tailwindcss/forms'),],
   theme: {
     extend: {
       fontFamily: {
